+++ modules/create_book.rb
@@ -1,0 +1,10 @@
+module CreateBook
+  def create_book
+    print 'Title: '
+    title = gets.chomp
+    print 'Author: '
+    author = gets.chomp
+    Book.new(title, author)
+    puts 'Book succesfully created!'
+  end
+end
