+++ modules/show_books.rb
@@ -1,0 +1,11 @@
+module ListBooks
+  def show_books
+    if Book.all.empty?
+      puts 'There\'s no books added yet!'
+    else
+      Book.all.each do |book|
+        puts %(Title: "#{book.title}", Author: #{book.author})
+      end
+    end
+  end
+end
